requirements.txt
openai>=1.0.0
PyMuPDF
python-docx
